Develop

+++ app/Http/Controllers/Admin/Order/OrderController.php
@@ -169,6 +169,156 @@ class OrderController extends Controller
         ->make(true);
     }
 
+    // public function dailyfoodorderlist(Request $request)
+    // {
+    //     if($request['start_date']){
+    //         $date_start=date('Y-m-d 00:00:00',strtotime($request['start_date']));
+    //     }else{
+    //         $date_start=date('Y-m-d 00:00:00');
+    //     }
+    //     if($request['end_date']){
+    //         $date_end=date('Y-m-d 23:59:59',strtotime($request['end_date']));
+    //     }else{
+    //         $date_end=date('Y-m-d 23:59:59');
+    //     }
+    //     $total_orders=CustomerOrder::whereBetween('created_at',[$date_start,$date_end])->where('order_type','food')->orderBy('order_id','desc')->paginate(15);
+    //     $filter_count=CustomerOrder::whereBetween('created_at',[$date_start,$date_end])->where('order_type','food')->count();
+    //     $all_count=CustomerOrder::where('order_type','food')->count();
+    //     $processing_orders=CustomerOrder::whereBetween('created_at',[$date_start,$date_end])->where('order_type','food')->whereIn('order_status_id',[3,4,5,6,10])->count();
+    //     $restaurant_cancel_orders=CustomerOrder::whereBetween('created_at',[$date_start,$date_end])->where('order_type','food')->where('order_status_id',2)->count();
+    //     $customer_cancel_orders=CustomerOrder::whereBetween('created_at',[$date_start,$date_end])->where('order_type','food')->where('order_status_id',9)->count();
+    //     $delivered_orders=CustomerOrder::whereBetween('created_at',[$date_start,$date_end])->where('order_type','food')->where('order_status_id',7)->count();
+    //     $pending_orders=CustomerOrder::whereBetween('created_at',[$date_start,$date_end])->where('order_type','food')->whereIn('order_status_id',[1,8])->count();
+
+    //     return view('admin.order.daily_food_orders.daily_index',compact('total_orders','all_count','filter_count','processing_orders','restaurant_cancel_orders','customer_cancel_orders','pending_orders','delivered_orders','date_start','date_end'));
+    // }
+
+    public function completeorderupdate(Request $request,$id)
+    {
+        $check_order=CustomerOrder::where('order_id',$id)->where('order_status_id',6)->first();
+        if($check_order){
+            CustomerOrder::where('order_id',$id)->update(['order_status_id'=>7,'is_admin_completed'=>1]);
+            NotiOrder::where('order_id',$id)->delete();
+            $check_order=CustomerOrder::where('order_id',$id)->first();
+            if($check_order->rider_id){
+                $has_order=CustomerOrder::where('rider_id',$check_order->rider_id)->whereIn('order_status_id',['3','4','5','6','10','12','13','14','17'])->first();
+                $check_rider=Rider::where('rider_id',$check_order->rider_id)->first();
+                    if($has_order){
+                        $check_rider->is_order=1;
+                        $check_rider->update();
+                    }else{
+                        $check_rider->is_order=0;
+                        $check_rider->update();
+                    }
+            }
+
+            $orderId=(string)$check_order->order_id;
+            $orderstatusId=(string)$check_order->order_status_id;
+            $orderType=(string)$check_order->order_type;
+            //rider
+            $rider_client = new Client();
+            $rider_token=$check_order->rider->rider_fcm_token;
+            if($rider_token){
+                $cus_url = "https://api.pushy.me/push?api_key=b7648d843f605cfafb0e911e5797b35fedee7506015629643488daba17720267";
+                try{
+                    $rider_client->post($cus_url,[
+                        'json' => [
+                            "to"=>$rider_token,
+                            "data"=> [
+                                "type"=> "rider_order_finished",
+                                "order_id"=>$orderId,
+                                "order_status_id"=>$orderstatusId,
+                                "order_type"=>$orderType,
+                                "title_mm"=> "Order Finished",
+                                "body_mm"=> "Good Day! Order is finished.Thanks very much!",
+                                "title_en"=> "Order Finished",
+                                "body_en"=> "Good Day! Order is finished.Thanks very much!",
+                                "title_ch"=> "订单已结束",
+                                "body_ch"=> "您的订单已结束! 再见！"
+                            ],
+                        ],
+                    ]);
+
+                }catch(ClientException $e){
+                }
+            }
+            //restaurant
+            $res_client = new Client();
+            if($check_order->restaurant->restaurant_fcm_token){
+                $res_token=$check_order->restaurant->restaurant_fcm_token;
+                $res_url = "https://api.pushy.me/push?api_key=67bfd013e958a88838428fb32f1f6ef1ab01c7a1d5da8073dc5c84b2c2f3c1d1";
+                try{
+                    $res_client->post($res_url,[
+                        'json' => [
+                            "to"=>$res_token,
+                            "data"=> [
+                                "type"=> "rider_order_finished",
+                                "order_id"=>$orderId,
+                                "order_status_id"=>$orderstatusId,
+                                "order_type"=>$orderType,
+                                "title_mm"=> "Order Finished",
+                                "body_mm"=> "Good Day! Order is finished.Thanks very much!",
+                                "title_en"=> "Order Finished",
+                                "body_en"=> "Good Day! Order is finished.Thanks very much!",
+                                "title_ch"=> "订单已结束",
+                                "body_ch"=> "订单已结束!",
+                                "sound" => "receiveNoti.caf",
+                            ],
+                            "mutable_content" => true ,
+                            "content_available" => true,
+                            "sound" => "receiveNoti.caf",
+                            "notification"=> [
+                                "title"=>"this is a title",
+                                "body"=>"this is a body",
+                                "sound" => "receiveNoti.caf",
+                            ],
+                        ],
+                    ]);
+
+                }catch(ClientException $e){
+                }
+            }
+            //customer
+            $cus_client = new Client();
+            if($check_order->customer->fcm_token){
+                $cus_token=$check_order->customer->fcm_token;
+                $cus_url = "https://api.pushy.me/push?api_key=cf7a01eccd1469d307d89eccdd7cee2f75ea0f588544f227c849a21075232d41";
+                try{
+                    $cus_client->post($cus_url,[
+                        'json' => [
+                            "to"=>$cus_token,
+                            "data"=> [
+                                "type"=> "rider_order_finished",
+                                "order_id"=>$orderId,
+                                "order_status_id"=>$orderstatusId,
+                                "order_type"=>$orderType,
+                                "title_mm"=> "Order Finished",
+                                "body_mm"=> "Good Day! Your order is finished. Thanks very much!",
+                                "title_en"=> "Order Finished",
+                                "body_en"=> "Good Day! Your order is finished. Thanks very much!",
+                                "title_ch"=> "订单已结束",
+                                "body_ch"=> "您的订单已结束! 祝您用餐愉快！再见！"
+                            ],
+                            "mutable_content" => true ,
+                            "content_available" => true,
+                            "notification"=> [
+                                "title"=>"this is a title",
+                                "body"=>"this is a body",
+                            ],
+                        ],
+                    ]);
+
+                }catch(ClientException $e){
+                }
+            }
+            $request->session()->flash('alert-success', 'successfully completed order!');
+            return redirect()->back();
+        }else{
+            $request->session()->flash('alert-warning', 'warning completed order! this order not exists in start delivery condation');
+            return redirect()->back();
+        }
+    }
+
     public function dailyfoodorderindex()
     {
         return view('admin.order.daily_food_orders.index');
@@ -261,10 +411,17 @@ class OrderController extends Controller
             }
             return $order_status;
         })
-        ->addColumn('action', function(CustomerOrder $post){
-            $view = '<a href="/fatty/main/admin/food_orders/view/'.$post->order_id.'" class="btn btn-primary btn-sm mr-2"><i class="fas fa-eye"></i></a>';
-            $pending = '<a href="/fatty/main/admin/pending/orders/define/'.$post->order_id.'" onclick="return confirm(\'Are You Sure Want to Pending Order?\')" class="btn btn-primary btn-sm mr-2"><i class="fas fa-plus-circle"></i></a>';
-            $btn=$view.$pending;
+        ->addColumn('detail', function(CustomerOrder $post){
+            $btn = '<a href="/fatty/main/admin/food_orders/view/'.$post->order_id.'" class="btn btn-info btn-sm mr-2" title="order detail"><i class="fas fa-eye"></i></a>';
+
+            return $btn;
+        })
+        ->addColumn('pending', function(CustomerOrder $post){
+            $btn = '<a href="/fatty/main/admin/pending/orders/define/'.$post->order_id.'" onclick="return confirm(\'Are You Sure Want to Pending Order?\')" class="btn btn-primary btn-sm mr-2" title="Order Pending"><i class="fas fa-plus-circle"></i></a>';
+            return $btn;
+        })
+        ->addColumn('complete', function(CustomerOrder $post){
+            $btn = '<a href="/fatty/main/admin/complete_order/update/'.$post->order_id.'" onclick="return confirm(\'Are You Sure Want to Complete Order?\')" class="btn btn-success btn-sm mr-2" title="Order Complete"><i class="fas fa-plus-circle"></i></a>';
 
             return $btn;
         })
@@ -282,7 +439,7 @@ class OrderController extends Controller
             }
             return $btn;
         })
-        ->rawColumns(['action','ordered_date','customer_type','status','payment_method_name'])
+        ->rawColumns(['pending','ordered_date','customer_type','status','payment_method_name','detail','complete'])
         ->searchPane('model', $model)
         ->make(true);
     }
@@ -916,22 +1073,28 @@ class OrderController extends Controller
     }
     public function pendingorderdefine(Request $request,$id)
     {
-        CustomerOrder::where('order_id',$id)->update(['order_status_id'=>8]);
-        NotiOrder::where('order_id',$id)->delete();
-        $check_order=CustomerOrder::where('order_id',$id)->first();
-        if($check_order->rider_id){
-            $has_order=CustomerOrder::where('rider_id',$check_order->rider_id)->whereIn('order_status_id',['3','4','5','6','10','12','13','14','17'])->first();
-            $check_rider=Rider::where('rider_id',$check_order->rider_id)->first();
-                if($has_order){
-                    $check_rider->is_order=1;
-                    $check_rider->update();
-                }else{
-                    $check_rider->is_order=0;
-                    $check_rider->update();
-                }
+        $check_order=CustomerOrder::where('order_id',$id)->where('order_status_id',6)->first();
+        if($check_order){
+            CustomerOrder::where('order_id',$id)->update(['order_status_id'=>8,'is_admin_completed'=>1]);
+            NotiOrder::where('order_id',$id)->delete();
+            $check_order=CustomerOrder::where('order_id',$id)->first();
+            if($check_order->rider_id){
+                $has_order=CustomerOrder::where('rider_id',$check_order->rider_id)->whereIn('order_status_id',['3','4','5','6','10','12','13','14','17'])->first();
+                $check_rider=Rider::where('rider_id',$check_order->rider_id)->first();
+                    if($has_order){
+                        $check_rider->is_order=1;
+                        $check_rider->update();
+                    }else{
+                        $check_rider->is_order=0;
+                        $check_rider->update();
+                    }
+            }
+            $request->session()->flash('alert-success', 'successfully Pending Order!');
+            return redirect()->back();
+        }else{
+            $request->session()->flash('alert-warning', 'warning pending order! this order not exists in start delivery condation');
+            return redirect()->back();
         }
-        $request->session()->flash('alert-success', 'successfully Pending Order!');
-        return redirect()->back();
     }
 
     public function assign_noti(Request $request,$id)
